moved use-actions hook
